Handle unused parameters in a systematic way
This change makes it possible to compile OpenSSL
with "-Wunused-parameters" without a warning. The
changes brings the code-base closer to a silent
compilation with "-Wextra".

In detail, the change follows the current best practice
of the OpenSSL code basis to flagged unused parameters with
"ossl_unused". This macro uses "__attribute__((unused))"
which can silence the compiler for a single occurrence
of a variable. However, the attributes means "possibly
unused", it does not complain, when an "ossl_unused"
attribute is actually used. Therefore, the change prefixes
all variable names with "UNUSED_" to detect incorrectly
flagged parameters as well. The change is fully portable
and works on all compilers.

For mixed cases (parameter is unused under certain conditions
such as e.g. debug or exclusion flags), the following idiom
was used on few occurrences.

  (void)buf;    /* silence -Wunused-parameter */