fixup! EVP: Preserve the EVP_PKEY id in a few more spots
EVP_PKEY_type() is dodgy!  It will translate "SM2" to EVP_PKEY_EC
because of aliasing, and "RSA" to the NID_rsa, which is obsolete...
furthermore, it dawned on me that we had already done this sort of
exercise, inside EVP_PKEY_is_a(), all that was needed was to extract
it into a function of its own.