Revert two renamings backported from master
The original names were more intuitive: the generate_counter counts the
number of generate requests, and the reseed_counter counts the number
of reseedings (of the principal DRBG).

    reseed_gen_counter  -> generate_counter
    reseed_prop_counter -> reseed_counter

This is the anologue to commit 8380f453ec81d9172b94a82c3592f78f1a612046
on the 1.1.1 stable branch. The only difference is that the second
renaming has already been reverted on the master branch.