provider: add input_type pem for RSA decoder
This commmit adds the input_type pem to the RSA decoder.

The motivation for this is an issue we ran into when upgrading Node.js
to OpenSSL 3.0. When reading a BIO, there will be a peek of the file to
see if it is of type PEM:

void *file_attach(void *provctx, OSSL_CORE_BIO *cin)
{
  ...
  peekbuf[sizeof(peekbuf) - 1] = '\0';
  if (strstr(peekbuf, "-----BEGIN ") != NULL)
    input_type = INPUT_TYPE_PEM;
  ...
  if (BIO_tell(new_bio) != loc) {
    /* In this case, anything goes */
    input_type = INPUT_TYPE_ANY;
}

When calling PEM_read_bio_PrivateKey with a BIO from a file, BIO_tell
would be supported and the input type is set to INPUT_TYPE_ANY (so we
don't hit this issue in that case). But in Node's case the input_type
will be "pem" and later when trying to find a decoder for that
input_type leading to the issue we are seeing.