Add RAND_swap_{primary,public,private}
These are internal functions that allow tests to switch to a
different RNG implementation, such as to do known-answer-tests.

Move the EVP_RAND declarations from crypto/evp files into crypto/rand.h
so they are internally available.

Implement a "fake" DRBG that fills in the buffer with fixed values, or
assigned values.  Tt's in the default provider and only accessible from
within OpenSSL. Convert ecdsatest, sm2test, and fuzz tests to use it.